pyproject.toml
[project]
name = "jobscanner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.14.0",
    "stagehand-py>=0.3.10",
]
